add home stack
