+++ src/navigation/UserStack.js
@@ -11,8 +11,12 @@ import { getAuth, signOut } from "firebase/auth";
 import CameraStack from "../navigation/CameraStack";
 import StoriesStack from "../navigation/StoriesStack";
 import ResourceStack from "../navigation/ResourceStack";
+import HomeStack from "./HomeStack";
 // import ChatStack from "../navigation/ChatStack";
 import { createStackNavigator } from "@react-navigation/stack";
+
+//Screens
+import HomeReadScreen from "../screens/HomeReadScreen";
 
 const Stack = createNativeStackNavigator();
 const Tab = createBottomTabNavigator();
@@ -32,30 +36,30 @@ const Tab = createBottomTabNavigator();
 //   }
 // })
 
-export default function UserStack() {
-  const auth = getAuth();
-  const user = auth.currentUser;
+export default function UserStack({navigation}) {
+  // const auth = getAuth();
+  // const user = auth.currentUser;
 
-  let screenOptions = {
-    headerShown: false,
-    tabBarShowLabel: false,
-    headerLeft: () => (
-      <Button
-        onPress={() => {
-          signOut(auth)
-            .then(() => {
-              // Sign-out successful.
-              user = null;
-            })
-            .catch((error) => {
-              // An error happened.
-              // should we do something with that error??
-            });
-        }}
-        title="Log Out"
-      />
-    ),
-  };
+  // let screenOptions = {
+  //   headerShown: false,
+  //   tabBarShowLabel: false,
+  //   headerLeft: () => (
+  //     <Button
+  //       onPress={() => {
+  //         signOut(auth)
+  //           .then(() => {
+  //             // Sign-out successful.
+  //             user = null;
+  //           })
+  //           .catch((error) => {
+  //             // An error happened.
+  //             // should we do something with that error??
+  //           });
+  //       }}
+  //       title="Log Out"
+  //     />
+  //   ),
+  // };
 
   return (
     <NavigationContainer>
